Big changes to GUI buttons. Created new working JFrame, porting in functionality soon.
