+++ Calculator/nbproject/project.properties
@@ -53,7 +53,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=calculator.ConvertToPostfix
+main.class=calculator.MainWindow
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
