--- Calculator/src/calculator/CalculatorWindow.java
@@ -1,300 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package calculator;
-import java.util.*;
-
-/**
- *
- * @author Dasmann
- */
-public class CalculatorWindow extends javax.swing.JPanel {
-
-    private ArrayList<String> exp = new ArrayList();
-    /**
-     * Creates new form CalculatorWindow
-     */
-    public CalculatorWindow() {
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        outputLabel = new javax.swing.JLabel();
-        clearButton = new javax.swing.JButton();
-        deleteButton = new javax.swing.JButton();
-        powButton = new javax.swing.JButton();
-        divButton = new javax.swing.JButton();
-        sevenButton = new javax.swing.JButton();
-        eightButton = new javax.swing.JButton();
-        nineButton = new javax.swing.JButton();
-        multButton = new javax.swing.JButton();
-        fourButton = new javax.swing.JButton();
-        fiveButton = new javax.swing.JButton();
-        sixButton = new javax.swing.JButton();
-        plusButton = new javax.swing.JButton();
-        oneButton = new javax.swing.JButton();
-        twoButton = new javax.swing.JButton();
-        threeButton = new javax.swing.JButton();
-        minusButton = new javax.swing.JButton();
-        zeroButton = new javax.swing.JButton();
-        leftParthButton = new javax.swing.JButton();
-        rightParthButton = new javax.swing.JButton();
-        equalsButton = new javax.swing.JButton();
-
-        outputLabel.setFont(new java.awt.Font("Tahoma", 0, 48)); // NOI18N
-        outputLabel.setHorizontalAlignment(javax.swing.SwingConstants.RIGHT);
-        outputLabel.setText("3*8+4-9");
-
-        clearButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        clearButton.setText("C");
-        clearButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        deleteButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        deleteButton.setText("DE");
-        deleteButton.setPreferredSize(new java.awt.Dimension(80, 60));
-        deleteButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                deleteButtonActionPerformed(evt);
-            }
-        });
-
-        powButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        powButton.setText("^");
-        powButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        divButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        divButton.setText("/");
-        divButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        sevenButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        sevenButton.setText("7");
-        sevenButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        eightButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        eightButton.setText("8");
-        eightButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        nineButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        nineButton.setText("9");
-        nineButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        multButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        multButton.setText("*");
-        multButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        fourButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        fourButton.setText("4");
-        fourButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        fiveButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        fiveButton.setText("5");
-        fiveButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        sixButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        sixButton.setText("6");
-        sixButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        plusButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        plusButton.setText("+");
-        plusButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        oneButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        oneButton.setText("1");
-        oneButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        twoButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        twoButton.setText("2");
-        twoButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        threeButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        threeButton.setText("3");
-        threeButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        minusButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        minusButton.setText("-");
-        minusButton.setPreferredSize(new java.awt.Dimension(80, 60));
-        minusButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                minusButtonActionPerformed(evt);
-            }
-        });
-
-        zeroButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        zeroButton.setText("0");
-        zeroButton.setPreferredSize(new java.awt.Dimension(80, 60));
-        zeroButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                zeroButtonActionPerformed(evt);
-            }
-        });
-
-        leftParthButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        leftParthButton.setText("(");
-        leftParthButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        rightParthButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        rightParthButton.setText(")");
-        rightParthButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        equalsButton.setFont(new java.awt.Font("Tahoma", 0, 36)); // NOI18N
-        equalsButton.setText("=");
-        equalsButton.setPreferredSize(new java.awt.Dimension(80, 60));
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
-        this.setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                        .addComponent(outputLabel, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addGroup(layout.createSequentialGroup()
-                            .addComponent(clearButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                            .addComponent(deleteButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                            .addComponent(powButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                            .addComponent(divButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                    .addGroup(layout.createSequentialGroup()
-                        .addComponent(sevenButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(eightButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(nineButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(multButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(layout.createSequentialGroup()
-                        .addComponent(fourButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(fiveButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(sixButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(plusButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(layout.createSequentialGroup()
-                        .addComponent(oneButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(twoButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(threeButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(minusButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(layout.createSequentialGroup()
-                        .addComponent(zeroButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(leftParthButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(rightParthButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(equalsButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(outputLabel, javax.swing.GroupLayout.PREFERRED_SIZE, 87, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(clearButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(deleteButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(powButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(divButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(sevenButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(eightButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(nineButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(multButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(fourButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(fiveButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(sixButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(plusButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(oneButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(twoButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(threeButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(minusButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(zeroButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(leftParthButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(rightParthButton, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(equalsButton, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        outputLabel.getAccessibleContext().setAccessibleName("");
-        deleteButton.getAccessibleContext().setAccessibleName("Delete");
-        powButton.getAccessibleContext().setAccessibleName("Power");
-        divButton.getAccessibleContext().setAccessibleName("Divide");
-        sevenButton.getAccessibleContext().setAccessibleName("Seven");
-        eightButton.getAccessibleContext().setAccessibleName("Eight");
-        nineButton.getAccessibleContext().setAccessibleName("Nine");
-        multButton.getAccessibleContext().setAccessibleName("Multiply");
-        fourButton.getAccessibleContext().setAccessibleName("Four");
-        fiveButton.getAccessibleContext().setAccessibleName("Five");
-        sixButton.getAccessibleContext().setAccessibleName("Six");
-        plusButton.getAccessibleContext().setAccessibleName("Plus");
-        oneButton.getAccessibleContext().setAccessibleName("One");
-        twoButton.getAccessibleContext().setAccessibleName("Two");
-        threeButton.getAccessibleContext().setAccessibleName("Three");
-        minusButton.getAccessibleContext().setAccessibleName("Minus");
-        zeroButton.getAccessibleContext().setAccessibleName("Zero");
-        leftParthButton.getAccessibleContext().setAccessibleName("LeftParenth");
-        rightParthButton.getAccessibleContext().setAccessibleName("RightParenth");
-        equalsButton.getAccessibleContext().setAccessibleName("Equals");
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void deleteButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteButtonActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_deleteButtonActionPerformed
-
-    private void minusButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_minusButtonActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_minusButtonActionPerformed
-
-    private void zeroButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_zeroButtonActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_zeroButtonActionPerformed
-
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton clearButton;
-    private javax.swing.JButton deleteButton;
-    private javax.swing.JButton divButton;
-    private javax.swing.JButton eightButton;
-    private javax.swing.JButton equalsButton;
-    private javax.swing.JButton fiveButton;
-    private javax.swing.JButton fourButton;
-    private javax.swing.JButton leftParthButton;
-    private javax.swing.JButton minusButton;
-    private javax.swing.JButton multButton;
-    private javax.swing.JButton nineButton;
-    private javax.swing.JButton oneButton;
-    private javax.swing.JLabel outputLabel;
-    private javax.swing.JButton plusButton;
-    private javax.swing.JButton powButton;
-    private javax.swing.JButton rightParthButton;
-    private javax.swing.JButton sevenButton;
-    private javax.swing.JButton sixButton;
-    private javax.swing.JButton threeButton;
-    private javax.swing.JButton twoButton;
-    private javax.swing.JButton zeroButton;
-    // End of variables declaration//GEN-END:variables
-}
